tests: remove error failure condition
Previously, some tests within the flash validation testsuite
were set to fail when the error prefix was detected. This is
now removed per instruction from the Ocre community developers.
Now, the flash validation testsuite includes only one testcase
which just checks for the expected hello world string and fails
if it is not detected.

Signed-off-by: Patrick Robb <probb@iol.unh.edu>