tests: Move test group execution to build workflow
Previously, the test groups were invoked within a dedicated tests
workflow, triggered by a workflow_run event. This did not work well
because workflows triggered by workflow_run are forced to run from the
main branch as a security measure. So, changes to the test workflow
coming in from a pull request were not used in the CI testing.

This commit also includes an increased sleep on the flash validation
test, allowing it to properly initialize the hello world container and
fill the output buffer with the expected string.

Signed-off-by: Patrick Robb <probb@iol.unh.edu>