tests: Add initial test case for Modbus server
Includes new test group for the modbus server, alongside the initial
test case where the modbus server is coneceted to by a client on the
github action's runner. Currently, the github actions runner connects to
the modbus server to read and write to the LED and Button register(s).
Will need to update test case as registers are implemented in the modbus
server container.

This commit also includes changes in tests.yml in order to run the test
group and store the github artifact.

Signed-off-by: Matthew Gee <mgee@iol.unh.edu>