tests: remove flash validation error failure condition (#83)
* tests: remove error failure condition
* tests: Move test group execution to build workflow

Now, the flash validation testsuite includes only one testcase
which just checks for the expected hello world string and fails
if it is not detected.

This commit also includes an increased sleep on the flash validation
test, allowing it to properly initialize the hello world container and
fill the output buffer with the expected string.

---------

Signed-off-by: Patrick Robb <probb@iol.unh.edu>