fix(sample_container): truncate file (#100)
If there is already a file in the filesystem with the same name
and it is larger than the new size we are writing, the file size
will not be truncated and we will end up with junk after the
end of the wasm buffer.

Use FS_O_TRUNC, so that the file will always be truncated
to the new size.

Signed-off-by: Marco Casaroli <marco.casaroli@gmail.com>