Github Artifact Cleanup
Removed artifacting of log files and combined wasm artifact uploads into
a singluar artifact. Currently the build wasm step saves all artifacts
to a directory local on disk and they are artifacted in a seperate step.
The directory used is cleaned at the end of the workflow to prevent runs
from affecting each other.

Signed-off-by: Matthew Gee <mgee@iol.unh.edu>