fix(shell): handle getopt properly
We use + in getopt to make sure gnu getopt will stop parsing
when it encounters non option arguments.

It looks like this does not hurt Zephyr.

We also parse the -v in ocre shell command to print the context
and set verbose mode, that can be used in the future.

Signed-off-by: Marco Casaroli <marco.casaroli@gmail.com>