feat(samples/mini): handle sample file
In the mini configuration, sometimes we will not be able to write
the state_information partition, so we just embed the file in the
binary and it will create it if it is non-existent.

The user can also override the image using the cmake variable
OCRE_INPUT_FILE.

We also ditch xxd and instead use a more portable solution using
awk, because xxd in Ubuntu 22.04 LTS does not support option -n.

Signed-off-by: Marco Casaroli <marco.casaroli@gmail.com>