tests: Add initial test case for Modbus Server Container (#108)
Includes new test group for the modbus server, alongside the initial
test case where the modbus server is coneceted to by a client on the
github action's runner. Currently, the github actions runner connects to
the modbus server to read and write to the LED and Button register(s).
Will need to update test case as registers are implemented in the modbus
server container.

This commit also includes a refactoring of parts of build.yml. Sets the
concurrency stategy for running multiple PR workflows to be in serial
rather than syncronous to prevent different runs affecting each other.
Added the ability for the github actions runner to flash the
b_u585i_iot02a board with the modbus server container. Also changes
overall workflow slightly in order to make future container testing on
the the b_u585i_iot02a board easier to implement.

---------

Signed-off-by: Matthew Gee <mgee@iol.unh.edu>