fix(container, wamr): race condition on container_kill
There was a race condition, where if we called
ocre_container_kill just after ocre_container_start (detached)
but before the thread gets created and cleared the WASM exception.

In this scenario, the kill command would be processed before the
cleaning of the exception, and ocre sets the status to stopped
but still the container starts later, yielding wrong behavior
if the containers is requested to be killed fast enough.

We change the semantics of the runtime engine interface, so we
pass a conditional variable that must be signalled by this the
thread_execute function, allowing the start function to
return just after the runtime engine can handle the kill command.

The documentation about the vtable will be updated regarding this
expectation.

Signed-off-by: Marco Casaroli <marco.casaroli@gmail.com>