tests: add system tests
System Tests by default handle Ocre Library as a black box.

By default, they build ocre in Release mode as we want to test
what we are delivering.

We can also reuse them as part of source coverage generation and
memory leak checks. This is why they should not tweak the build
time variables, and allow us to override them, in case we want
to use these tests.

Signed-off-by: Marco Casaroli <marco.casaroli@gmail.com>